server: add long options, use env var as bind path, add protocol check
server.c:
- Replace getopt with getopt_long, add long options for existing ones.
- Add -u or --unix option with an optional argument to specify the use
of Unix socket. If no argument is specified, check whether the
UNIX_SOCKET environmental variable is set, use its value as path
to bind the socket to, if not set use the default path UNIX_SOCKET_PATH
specified in pingpong.h. Otherwise, use the argument as path to bind to.
- Add protocol enum, and variable which is set during the option parsing
to specified protocol. If no protocol is specified exit with error.
- Update usage message to accomodate long options

pingpong.h.in:
- Update handle_error comment
- Add the UNIX_SOCKET_PATH macro, which is the default path to bind the
Unix socket to if no path was specified explicitly.