server: add tcp ping responder, update option parser
server.c:
- Remove protocol enum, use sock_domain instead.
- Update Unix socket argument parser. Initialize unix_sock_path only
there.
- Add --listen|-L options, and update usage message correspondingly.
- Add Inet socket argument parser.
- Add check for conflicting protocols in parse_options.
- Redirect error output prints to stderr.
- Add socket domain switch to main. Since socket creation and binding
process is different for inet and unix sockets, the switch is used.
- Move unix socket creation, binding, and address structure
initialization to AF_UNIX case of domain switch.
- Add inet socket creation, binding, and address selection to AF_INET
case of domain switch. Address selection is performed with the help of
getaddrinfo, and the process is as described in getaddrinfo manpage.
- Replace read with recv, and write with send for operations on sockets.
This is useful in case we want to set flags. However, no flags are used.
- Unlink unix_sock_path on exit only if it was set beforehand.

pingpong.h.in:
- Add defines for default inet socket address and port.