client: add tcp ping sender, update option parser
client.c:
- Remove protocol enum, use sock_domain instead.
- Update Unix socket argument parser. Initialize unix_sock_path only
there.
- Add --connect|-C options, and update usage message correspondingly.
- Add Inet socket argument parser.
- Add check for conflicting protocols in parse_options.
- Redirect error output prints to stderr.
- Add socket domain switch to main. Since socket creation and
connection process is different for inet and unix sockets,
the switch is used.
- Move unix socket creation, connection, and address structure
initialization to AF_UNIX case of domain switch.
- Add inet socket creation, connection, and address selection to AF_INET
case of domain switch. Address selection is performed with the help of
getaddrinfo, and the process is as described in getaddrinfo manpage.
- Replace read with recv, and write with send for operations on sockets.
This is useful in case we want to set flags. However, no flags are used.