docker: add Dockerfile, add "run" shell script
This version is concerned with running the Unix socket versions of
server and client in the same container. It, however, for some reason
does not produce any output when container is run with docker run
command.

Dockerfile: Add two-stage dockerfile, corresponding images are Build,
where the compilation is done, and Run which contains server and client
binaries and the run script.

run.sh: add bash script which calls the server and the client binaries
in Unix socket mode, with path argument specified in the environmental
variable. The script detaches each process from the shell. This is due
to the server blocking until the client sends a message, and the client
must be ran after the server.