cmake, docker: link log.c library as shared library
Top-level CMakeLists.txt:
- Add log.c library as shared instead of static. There is no particular
reason for using log.c library as shared library, it is even better to
use it as a static library, as it would decrease the size of the built
program.
- Set CMAKE_LIBRARY_OUTPUT_DIRECTORY to CMAKE_BINARY_DIR/lib. This is
useful when building docker image, as the shared object files can be
easliy found there.

Dockerfile:
- Add copy build/lib contents to /usr/lib, to allow the ld.so
linker/loader find the log.c shared library.